[WPT] Prefetch + ServiceWorker: Check FetchEvent
This CL adds and checks `interceptedRequests` that captures the
FetchEvents intercepted by the ServiceWorker, namely its headers.

To check the headers from ServiceWorker FetchEvent (from
https://fetch.spec.whatwg.org/#headers) in this CL and the headers from
WPT server (in `executor.sub.html`) in the same way, preceding CLs
https://chromium-review.googlesource.com/c/chromium/src/+/6363856 etc.
have changed the format of the header information.

This CL makes the tests with modifying requests when ServiceWorker
support is not enabled, by detecting the request before modification is
not a prefetch request.

Bug: 40947546
Change-Id: I71456475a6a7bde437f640aa648ede522ed9b681